move tests #5622
